Rename Open Source Registry to Distribution
