+++ internal/events/distribution/types.go
@@ -1,4 +1,4 @@
-package osr
+package distribution
 
 import (
 	"context"
