+++ kobold/config/config.go
@@ -32,10 +32,10 @@ type CommitMessageSpec struct {
 type EndpointType string
 
 const (
-	EndpointTypeGeneric   EndpointType = "generic"
-	EndpointTypeACR       EndpointType = "acr"
-	EndpointTypeDockerhub EndpointType = "dockerhub"
-	EndpointTypeOSR       EndpointType = "osr"
+	EndpointTypeGeneric      EndpointType = "generic"
+	EndpointTypeACR          EndpointType = "acr"
+	EndpointTypeDockerhub    EndpointType = "dockerhub"
+	EndpointTypeDistribution EndpointType = "distribution"
 )
 
 type Header struct {
